requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
discord.py==1.6.0
gTTS==2.2.2
ibm-cloud-sdk-core==3.8.0
ibm-watson==5.1.0
idna==2.10
jaconv==0.2.4
mecab==0.996.2
multidict==5.1.0
pycparser==2.20
PyJWT==2.0.1
pymemcache==3.4.1
PyNaCl==1.4.0
python-dateutil==2.8.1
redis==3.5.3
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.4
websocket-client==0.48.0
yarl==1.6.3
